Clean up
- Remove unused `ToolHelper` methods.
- Move measurement rendering into `AbstractDrawingLikeTool` so it isn't duplicated.
- Rename `AbstractDrawingTool` to `AbstractTemplateTool` since that is all it does now. Also remove anything that isn't
  used anymore.
