+++ src/main/java/net/rptools/maptool/client/tool/drawing/DrawingTool.java
@@ -25,9 +25,7 @@ import javax.annotation.Nullable;
 import javax.swing.SwingUtilities;
 import net.rptools.maptool.client.MapTool;
 import net.rptools.maptool.client.MapToolUtil;
-import net.rptools.maptool.client.ScreenPoint;
 import net.rptools.maptool.client.swing.colorpicker.ColorPicker;
-import net.rptools.maptool.client.tool.ToolHelper;
 import net.rptools.maptool.client.ui.zone.renderer.ZoneRenderer;
 import net.rptools.maptool.model.Zone;
 import net.rptools.maptool.model.ZonePoint;
@@ -226,39 +224,7 @@ public final class DrawingTool<StateT> extends AbstractDrawingLikeTool {
         drawable.draw(renderer.getZone(), g2, pen);
 
         // Measurements
-        var measurement = result.measurement();
-        switch (measurement) {
-          case null -> {}
-          case Measurement.Rectangular rectangular -> {
-            var rectangle = rectangular.bounds();
-            ToolHelper.drawBoxedMeasurement(
-                renderer,
-                g,
-                ScreenPoint.fromZonePoint(renderer, rectangle.getX(), rectangle.getY()),
-                ScreenPoint.fromZonePoint(renderer, rectangle.getMaxX(), rectangle.getMaxY()));
-          }
-          case Measurement.LineSegment lineSegment -> {
-            var p1 =
-                ScreenPoint.fromZonePoint(
-                    renderer, lineSegment.p1().getX(), lineSegment.p1().getY());
-            var p2 =
-                ScreenPoint.fromZonePoint(
-                    renderer, lineSegment.p2().getX(), lineSegment.p2().getY());
-            ToolHelper.drawMeasurement(renderer, g, p1, p2);
-          }
-          case Measurement.IsoRectangular isoRectangular -> {
-            var north =
-                ScreenPoint.fromZonePoint(
-                    renderer, isoRectangular.north().getX(), isoRectangular.north().getY());
-            var west =
-                ScreenPoint.fromZonePoint(
-                    renderer, isoRectangular.west().getX(), isoRectangular.west().getY());
-            var east =
-                ScreenPoint.fromZonePoint(
-                    renderer, isoRectangular.east().getX(), isoRectangular.east().getY());
-            ToolHelper.drawIsoRectangleMeasurement(renderer, g, north, west, east);
-          }
-        }
+        drawMeasurementOverlay(renderer, g, result.measurement());
       }
     }
